Function renaming and some matching

+++ src/codeseg2/codeseg2_398.c
@@ -4,10 +4,10 @@
 struct unkfunc_80091060
 {
     u8 padding[0x38];
-    s32 unk38;
+    f32 unk38;
 };
 
-void func_80091060(struct unkfunc_80091060 *arg0, s32 arg1)
+void func_80091060(struct unkfunc_80091060 *arg0, f32 arg1)
 {
     arg0->unk38 = arg1;
 }
